Changed how time steps work, and added elapsed time statistics

+++ Prototypes/OrbitSim/src/sim/simobject/ObjectInSpace.java
@@ -45,7 +45,7 @@ public abstract class ObjectInSpace implements SimObject, GUIObject {
 	 * 
 	 * @param t the time elapsed in the step in seconds
 	 */
-	public void step(int t) {
+	public void step(long t) {
 		
 	}
